+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `--zip-single-file` and `--compression` options
 - Add progress bars for operations
 - Move to rusty-s3
+- Move to ulid to support purging expired files
 
 # 0.1.1